Remove Extra Scrollbars on Player Pane
There were too many nested divs with "overflow: scroll" on the player
pane, resulting in nested scrollbars. As a fix, I've removed this
property from the outermost #player div. I chose the outermost div for
parity's sake with the editor pane, which puts the scrollbars on the
inner content rather than the outer div.

Fixes issue #66.