Use Cached Suggestion Values
Rather than potentially re-parsing every document in the project on every keypress for suggestion purposes, this change will allow InkFileSymbols's existing asynchronous dirty parsing to handle suggestion parsing. The results are cached for use by InkCompleter when it needs them in the future.

This does introduce the possibility of the suggestions being slightly out-of-date, but the writer would have to maintain ~5 keystrokes per second for the entire time between the first use of a term and the second use in order for them to notice.