Zoom support for Inky!
Creates new menu options under View, "Zoom (Increase)" and "Zoom (Decrease)". Clicking either calls ProjectWindow.zoom() which, in turn, sends a IPC message to the renderer window on the event 'zoom' with the amount to zoom.

In the renderer window (Controller.js), this message is watched for and the reaction is to look for and change the fontSize properties of the elements with the id "#editor" and that of the id "#player".

Using accelerator settings, the text size can be zoomed through CmdOrCtrl+K, to increase, or CmdOrCtrl+L to decrease. The preferred options would have been to use CmdOrCtrl+Plus and CmdOrCtrl+Minus, but Electron does not have "Minus" as an available key code. (See https://electronjs.org/docs/api/accelerator).