Tentative fix for scroll issues in the web player
See #674. When playing a story in the web player, images are added to the div with a height of zero, which is updated after they load in. If they're loaded after the inner div's height is locked, the player won't be able to scroll down, potentially causing a soft lock. This fix locks the story's height after a choice is chosen, and unlocks it right before scrolling down. I'm not sure if it's a perfect solution, but it should prevent both soft locks and the jump upwards described in main.js circa line 193.